[wip] Page only when web components signal ready
Only paginate when child web components signal they are "ready" i.e.
have rendered so that Paged.js can accurately get their size.

This assumes that Paged.js is looking at the original size of elements
to correct paginate them.

It also relies on a mechanism by which web components can report that
rendering has completed, which does not exist for lit-element. This
would have to be patched into upstream, or a mixin developed.

This patch is hard-coded to work with <html-import>. A more general
solution would need to detect all compatible child custom elements and
wait for to receive readiness signal before paginating, or until some
time out.